--- a/apidoc/lib/jsduck_generator.js
+++ b/apidoc/lib/jsduck_generator.js
@@ -432,6 +432,7 @@ function exportAPIs(api, type) {
 exports.exportData = function exportJsDuck(apis) {
 	const rv = [];
 	doc = apis; // TODO make doc a field on a type, rather than this weird file-global!
+	common.createMarkdown(doc);
 
 	common.log(common.LOG_INFO, 'Annotating JSDuck-specific attributes...');