--- a/package.json
+++ b/package.json
@@ -116,8 +116,8 @@
     "husky": "^1.3.1",
     "js-yaml": "^3.12.1",
     "lint-staged": "^8.1.0",
+    "markdown-it": "^8.4.2",
     "mocha": "^5.2.0",
-    "pagedown": "~1.1.0",
     "ssri": "^6.0.1"
   },
   "repository": {