JIRA: https://jira.appcelerator.org/browse/TIDOC-3279

@jawa9000 this is a first try at moving to a new markdown parser to support fenced code blocks. I compared the output of the HTML generator and apart from a few newlines and some characters that are now escaped (like quotes and angle brackets) it's identical.

I'm not familiar with the complete process for generating the docs so it would be awesome if you can take another look and see if there are any issues i might have missed. For example, i'm not sure how the code highlighting was done in the old one. markdown-it has the ability to use a custom code highlighter via the `highlight ` option.